+++ include/TServer.h
@@ -8,6 +8,8 @@
 #include <mutex>
 #include <unordered_set>
 
+#include "BoostAliases.h"
+
 class TClient;
 class TNetwork;
 class TPPSMonitor;
@@ -19,19 +21,22 @@ public:
     TServer(const std::vector<std::string_view>& Arguments);
 
     void InsertClient(const std::shared_ptr<TClient>& Ptr);
-    std::weak_ptr<TClient> InsertNewClient();
     void RemoveClient(const std::weak_ptr<TClient>&);
     // in Fn, return true to continue, return false to break
     void ForEachClient(const std::function<bool(std::weak_ptr<TClient>)>& Fn);
     size_t ClientCount() const;
 
-    static void GlobalParser(const std::weak_ptr<TClient>& Client, std::string Packet, TPPSMonitor& PPSMonitor, TNetwork& Network);
+    static void GlobalParser(const std::weak_ptr<TClient>& Client, std::vector<uint8_t>&& Packet, TPPSMonitor& PPSMonitor, TNetwork& Network);
     static void HandleEvent(TClient& c, const std::string& Data);
     RWMutex& GetClientMutex() const { return mClientsMutex; }
 
-    
     const TScopedTimer UptimeTimer;
+
+    // asio io context
+    io_context& IoCtx() { return mIoCtx; }
+
 private:
+    io_context mIoCtx {};
     TClientSet mClients;
     mutable RWMutex mClientsMutex;
     static void ParseVehicle(TClient& c, const std::string& Pckt, TNetwork& Network);
@@ -39,4 +44,12 @@ private:
     static bool IsUnicycle(TClient& c, const std::string& CarJson);
     static void Apply(TClient& c, int VID, const std::string& pckt);
     static void HandlePosition(TClient& c, const std::string& Packet);
+};
+
+struct BufferView {
+    uint8_t* Data { nullptr };
+    size_t Size { 0 };
+    const uint8_t* data() const { return Data; }
+    uint8_t* data() { return Data; }
+    size_t size() const { return Size; }
 };