+++ src/THeartbeatThread.cpp
@@ -42,8 +42,6 @@ void THeartbeatThread::operator()() {
 
         Body += "&pps=" + Application::PPS();
 
-        beammp_trace("heartbeat body: '" + Body + "'");
-
         auto SentryReportError = [&](const std::string& transaction, int status) {
             auto Lock = Sentry.CreateExclusiveContext();
             Sentry.SetContext("heartbeat",
@@ -61,11 +59,10 @@ void THeartbeatThread::operator()() {
         bool Ok = false;
         for (const auto& Url : Application::GetBackendUrlsInOrder()) {
             T = Http::POST(Url, 443, Target, Body, "application/x-www-form-urlencoded", &ResponseCode, { { "api-v", "2" } });
-            beammp_trace(T);
             Doc.Parse(T.data(), T.size());
             if (Doc.HasParseError() || !Doc.IsObject()) {
                 if (!Application::Settings.Private) {
-                    beammp_error("Backend response failed to parse as valid json");
+                    beammp_trace("Backend response failed to parse as valid json");
                     beammp_trace("Response was: `" + T + "`");
                 }
                 Sentry.SetContext("JSON Response", { { "reponse", T } });
@@ -148,7 +145,7 @@ std::string THeartbeatThread::GenerateCall() {
         << "&map=" << Application::Settings.MapName
         << "&private=" << (Application::Settings.Private ? "true" : "false")
         << "&version=" << Application::ServerVersionString()
-        << "&clientversion=" << Application::ClientVersionString()
+        << "&clientversion=" << std::to_string(Application::ClientMajorVersion()) + ".0" // FIXME: Wtf.
         << "&name=" << Application::Settings.ServerName
         << "&modlist=" << mResourceManager.TrimmedList()
         << "&modstotalsize=" << mResourceManager.MaxModSize()